+++ __tests__/src/test-utils/configs.ts
@@ -22,6 +22,8 @@ export const getConfigs = (port: number) => {
       dbTable: 'inbox',
       postgresPub: 'pg_transactional_inbox_tests_pub',
       postgresSlot: 'pg_transactional_inbox_tests_slot',
+      restartDelay: 1,
+      maxRetries: 5,
     },
   };